dom/tests/mochitest/cowl/test_xhr_receive_labeled_object.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
  <script class="testbody" type="application/javascript">

    SimpleTest.waitForExplicitFinish();

    var completedTests = 0;
    var passedTests = 0;

    function ok_wrapper(result, desc) {
      ok(result, desc);

      completedTests++;

      if (result) {
        passedTests++;
      }

      if (completedTests == 10) {
        is(passedTests, completedTests, "There are " + completedTests + " privilege tests that should pass");

        SimpleTest.finish();
      }
    }

  </script>
  <iframe src="http://example.com/tests/dom/tests/mochitest/cowl/file_receive_labeled_object.html" frameborder="0"></iframe>
</pre>
</body>
</html>
